Fix big in mapped file. Fix integer overflow uint32_t -> uint64_t when playing large video files

+++ rfcommon/include/rfcommon/VideoEmbed.hpp
@@ -13,16 +13,16 @@ class MappedFile;
 class RFCOMMON_PUBLIC_API VideoEmbed : public RefCounted
 {
 public:
-    VideoEmbed(MappedFile* file, const void* data, uint32_t size);
+    VideoEmbed(MappedFile* file, const void* data, uint64_t size);
     virtual ~VideoEmbed();
 
     const void* address() const { return address_; }
-    uint32_t size() const { return size_; }
+    uint64_t size() const { return size_; }
 
 private:
     Reference<MappedFile> file_;
     const void* address_;
-    uint32_t size_;
+    uint64_t size_;
 };
 
 }
