+++ rfcommon/src/VideoEmbed.cpp
@@ -4,7 +4,7 @@
 namespace rfcommon {
 
 // ----------------------------------------------------------------------------
-VideoEmbed::VideoEmbed(MappedFile* file, const void* data, uint32_t size)
+VideoEmbed::VideoEmbed(MappedFile* file, const void* data, uint64_t size)
     : file_(file)
     , address_(data)
     , size_(size)
